+++ app/models/service.rb
@@ -12,6 +12,8 @@ class Service < ApplicationRecord
   has_many :service_categories, dependent: :destroy
   has_many :categories, through: :service_categories
   has_many :service_opinions, through: :project_items
+  has_many :service_research_areas, dependent: :destroy
+  has_many :research_areas, through: :service_research_areas
 
   has_many :source_relationships,
            class_name: "ServiceRelationship",
@@ -51,6 +53,7 @@ class Service < ApplicationRecord
   validates :restrictions, presence: true
   validates :phase, presence: true
   validates :logo, blob: { content_type: :image }
+  validates :research_areas, presence: true
 
   after_save :set_first_category_as_main!, if: :main_category_missing?