+++ app/models/service_research_area.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class ServiceResearchArea < ApplicationRecord
+  belongs_to :service
+  belongs_to :research_area
+
+  validates :service, presence: true
+  validates :research_area, presence: true, uniqueness: { scope: :service_id }
+end