+++ spec/models/service_spec.rb
@@ -28,6 +28,7 @@ RSpec.describe Service do
   it { should have_many(:service_categories).dependent(:destroy) }
   it { should have_many(:offers).dependent(:restrict_with_error) }
   it { should have_many(:categories) }
+  it { should have_many(:service_research_areas).dependent(:destroy) }
 
   it "sets first category as default" do
     c1, c2 = create_list(:category, 2)